\name{dpendensity}
\alias{dpendensity}
\alias{ppendensity}
\title{Calculating the fitted density or distribution}
\description{
 Calculating the fitted density or distribution. 
}
\usage{
dpendensity(x,val)
ppendensity(x,val)
}

\arguments{
	\item{x}{Object of class pendensity}
        \item{val}{Vector of values where to calculate the density}
}

\references{Density Estimation with a Penalized Mixture Approach, Schellhase C. and Kauermann G. (2012), Computational Statistics 27 (4), p. 757-777.}
\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}

\keyword{math}

