\name{print.pendensity}
\Rdversion{1.1}
\alias{print.pendensity}
\alias{print}

\title{Printing the main results of the (conditional) penalized density estimation}

\description{Printing the call of the estimation, the resulting weights, the final lambda0 and the corresponding value of AIC. Need an object of class pendensity.}

\usage{
\method{print}{pendensity}(x, ...)
}

\arguments{
  \item{x}{x has to be object of class pendensity}
  \item{...}{further arguments}
}

\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}

\keyword{print}
